requirements.txt
chardet==5.2.0
colorama==0.4.6
pillow==11.0.0
pyzbar==0.1.9
qrcode==8.0
reportlab==4.2.5
tk==0.1.0
